+++ components/NumberSpinner/NumberSpinner.js
@@ -176,6 +176,10 @@ class NumberSpinner extends Field {
     this.input.enable()
   }
 
+  _onBlur() {
+    this._callHandler(this.props.onBlur)
+  }
+
   _handleRules() {
     const { precision } = this.props
     const rules = []